refactor the error code
